Use different names to convey thread status
The `RUBY_INTERNAL_THREAD_EVENT_*` names convey state transition not status.
Given that `gvl-tracing` generate areas with a state until the next event,
it would be more clear to disaply thread status instead.

I think `running` is best to describe threads that have acquire the GVL
and are actually executing.

For `waiting` is the best I could find to describe thread waiting for IO or join.
But `sleeping` could do as well.

And finally I think `ready` works to describe threads that could be executed
but are waiting on the GVL.

Either way the README should probably describe the signification of each
status.