Rename ready -> wants_gvl and exited -> died
"ready" may not be clear it's usually not a great state for a thread
to be in for a long time, so I've renamed it to "wants_gvl" to
hopefully make it sound worse (or at least, not harmless).

The exact string was also chosen to get a nice contrasting color
on the perfetto UI.

I've also renamed "exited" to "died", since it seems more natural
to talk about threads that die vs exit.