Provide when possible the thread name and gem name
gvl-tracing handles properly thread name for basic usage like:
```ruby
  Thread.new do
    Thread.current.name = "thread name
  end
```

But on more complicated with external libraries it is not able to
properly handles thread name.
Jean had a good idea in https://github.com/ivoanjo/gvl-tracing/issues/9
about using `Thread.list`.

This commit does this:
- Store as list of json line, events we want to monitor
- Store `Thread.list` before stopping saving event related to the GVL
- Iterate of events store in json file and modify entries where we can
  find the name of the thread
- Go a little bit further when we can provide the gem name

This is a very naive approach and probably not super efficient. But this
is a first try.