Introduce abstraction for thread identifiers, and use serial counter for 3.3
To avoid having lots of if 3.3/else 3.2 code, I've introduced an
abstraction to get the thread identifier, which will isolate this
kind of code.

Also, I've chosen to use the thread serial as the thread identifier for
3.3 instead of the VALUE with some bits masked. We moved away from
thread serial in the past when we added the code to better fetch
thread names, but I think for 3.3 we can support it.

(This change doesn't solve thread names for 3.3 -- will come in a
later commit.)