Refactor state initialization code to actually run at... state initialization
Prior to this change, the initialization was only being done lazily,
when an event was rendered for a thread. This meant that, for instance,
on Ruby 3.3, if we asked for the serial of a thread that never got
initialized, we'd get a 0 instead of a serial (because the thread
hadn't been initialized).

Now the context is correctly initialized and the above can't happen.