Make thread names work even for dead threads for Ruby 3.3+
For a long time we've had a limitation on thread names, where we could
(sometimes) get the names of threads if they had been created before
GvlTracing started OR if they were alive then GvlTracing stopped.

But if a thread was created and died during the GvlTracing execution,
you were out of luck.

This has been annoying me for a long time, so here's a stab at fixing
it which seems to work and I believe is correct even in the presence of
Ractors.

The key idea is that we keep an array (`all_seen_threads`) that
contains references to every thread that we've ever seen. Then, when
stopping, we flush the names of all those threads to the trace file.

There's a few subtleties:

1. If a thread dies during GvlTracing, we still want to keep a
   reference to it around, since we want to render its name from Ruby
   code.

   But we don't want to create a `Thread` object memory leak.
   So, in `stop`, after we flush the names of the threads, we trim
   the array, keeping only alive threads in it -- so that further
   calls to `start` still can use them.

2. During `start` we also trim the alive threads, just in case a
   few more died after the last `stop`.

3. Because the `all_seen_threads` array is mutable global state, we
   need to be very careful because Ractors are a thing. Thus, we
   protect it using a mutex. From my understanding of the Ruby VM
   implementation and Ractor implementation, this is fine.
   (Please do let me know if I missed something).

4. All of this depends on us getting the `VALUE` of a Ruby thread,
   which we get on 3.3 but not on 3.2. So on 3.2, none of the above
   applies, and we fall back to the previous behavior of using
   `Thread.list` to render the names of whatever threads we still
   find there.

5. I'm using the C11 mutex APIs. I think anyone using GvlTracing
   is probably using a C11-capable compiler, but let me know if
   I'm mistaken.