Add lock to protect output from tearing
Got reports of tearing of output on macos, so let's see if this lock
is enough to solve the issue. I don't expect a big impact on
performance since we're talking about a few hundreds of events per
second, most of them not being concurrent.

In the future if this proves to be a bottleneck we can always evolve
into something way fancier.