Replace `rb_fiber_alive_p` with `Thread#alive?` via direct-bind
A thread's fiber get allocated lazily. In CI I spotted a few crashes
due to our call to `rb_fiber_alive_p` triggering allocations.

To fix this, we replace it with a call to `Thread#alive?` via the
direct-bind gem (as otherwise we couldn't directly call
`rb_thread_alive_p` as it's not a public symbol published by th VM).

This commit is also intended to act as an example for onboarding
direct-bind into a gem.